+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ### Removed
 - Removed the unused `saveModifiedBlocksOnly` setting
+- Removed `benf\neo\Field::$localizeBlocks`; use `$propagationMethod` instead
+- Removed `benf\neo\Field::$wasModified`
 - Removed `benf\neo\integrations\fieldlabels\FieldLabels`
 - Removed `benf\neo\converters\Field` (Neo Field Converter for Schematic)
 - Removed `benf\neo\converters\models\BlockType` (Neo BlockType Converter for Schematic)